WIP moved request execute to cursor()
